+++ jest.config.mjs
@@ -8,6 +8,7 @@ export default {
   testRegex: 'tests/.*\\.test\\.ts$',
   moduleNameMapper: {
     '^obsidian$': '<rootDir>/tests/__mocks__/obsidian.ts',
+    '\\.svg$': '<rootDir>/tests/__mocks__/svg-mock.ts',
   },
   collectCoverageFrom: [
     'src/logic/**/*.ts',